requirements.txt
fastapi
uvicorn
requests
sentence-transformers
faiss-cpu
numpy
huggingface_hub
PyMuPDF
spacy